+++ src/api2/api-caller.js
@@ -29,9 +29,11 @@ export class ApiCaller {
    * @param {Wallet} [opts.wallet] - the initialized {@link Wallet} instance for signing requests/transactions
    * @param {string} [opts.passphrase] - the passphrase of current TokenD network (is used internally when signing transactions)
    */
-  constructor (opts) {
-    this._axios = opts.axios
-    this._baseURL = opts.baseURL
+  constructor (opts = {}) {
+    this._axios = axios.create()
+    if (opts.baseURL) {
+      this.useBaseURL(opts.baseURL)
+    }
 
     this._wallet = null
     this._networkDetails = {}
@@ -47,10 +49,15 @@ export class ApiCaller {
     }
   }
 
+  withWallet (wallet) {
+    const newCaller = Object.assign(Object.create(Object.getPrototypeOf(this)), this)
+    newCaller.useWallet(wallet)
+    return newCaller
+  }
+
   static getInstance (baseURL) {
     return new ApiCaller({
-      baseURL,
-      axios: axios.create()
+      baseURL
     })
   }
 
@@ -276,5 +283,14 @@ export class ApiCaller {
 
   usePassphrase (networkPassphrase) {
     Network.use(new Network(networkPassphrase))
+  }
+
+  useBaseURL (baseURL) {
+    this._baseURL = baseURL
+  }
+
+  useNetworkDetails (networkDetails) {
+    this._networkDetails = networkDetails
+    this.usePassphrase(networkDetails.networkPassphrase)
   }
 }